Final-implementation/complete/requirements.txt
fastapi
uvicorn
requests
httpx
python-dotenv
python-multipart
openai